Ensure that Node.js polyfills are pre-optimized before the first request (#8688)
* Ensure that Node.js polyfills are pre-optimized before the first request

Previously, these polyfills were only optimized on demand when Vite became aware of them.
This was either because Vite was able to find an import to a polyfill when statically analysing the import tree of the entry-point,
or when a polyfilled module was dynamically imported as part of a executing code to handle a request.

In the second case, the optimizing of the dynamically imported dependency causes a reload of the Vite server, which can break applications that are holding state in modules during the request.
This is the case of most React type frameworks, in particular React Router.

Now, we pre-optimize all the possible Node.js polyfills when the server starts before the first request is handled.

* fixup! Ensure that Node.js polyfills are pre-optimized before the first request

* add test

* simplify the  nodejs warnings plugin

* fixup! Ensure that Node.js polyfills are pre-optimized before the first request

* fixup! simplify the  nodejs warnings plugin

* add debugging to e2e tests

* fixup! Ensure that Node.js polyfills are pre-optimized before the first request

* Initialize nodeJsCompatWarnings in configResolved hook

* Fix wrangler-configs-validation e2e tests

---------

+++ packages/vite-plugin-cloudflare/e2e/global-setup.ts
@@ -3,7 +3,7 @@ import os from "node:os";
 import path from "node:path";
 import util from "node:util";
 import { startMockNpmRegistry } from "@cloudflare/mock-npm-registry";
-import type { GlobalSetupContext } from "vitest/node";
+import type { TestProject } from "vitest/node";
 
 declare module "vitest" {
 	export interface ProvidedContext {
@@ -13,8 +13,7 @@ declare module "vitest" {
 
 // Using a global setup means we can modify tests without having to re-install
 // packages into our temporary directory
-// Typings for the GlobalSetupContext are augmented in `global-setup.d.ts`.
-export default async function ({ provide }: GlobalSetupContext) {
+export default async function ({ provide }: TestProject) {
 	const stopMockNpmRegistry = await startMockNpmRegistry(
 		"@cloudflare/vite-plugin"
 	);
@@ -28,7 +27,9 @@ export default async function ({ provide }: GlobalSetupContext) {
 	return async () => {
 		await stopMockNpmRegistry();
 
-		console.log("Cleaning up temporary directory...");
-		await fs.rm(root, { recursive: true, maxRetries: 10 });
+		if (!process.env.CLOUDFLARE_VITE_E2E_KEEP_TEMP_DIRS) {
+			console.log("Cleaning up temporary directory...", root);
+			await fs.rm(root, { recursive: true, maxRetries: 10 });
+		}
 	};
 }
